refactor(efi): own filesystem state in the SimpleFileSystem module
The mounted FAT geometry and the block device behind it are only touched by
the SimpleFileSystem protocol. Move them there and drop the unused
`FilesystemState::empty()`.
